app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "langchain-groq==0.1.0",
    "langgraph==0.0.40",
    "pydantic==2.5.0",
    "pyjwt==2.8.0",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "uvicorn[standard]==0.24.0",
]
